Refactor configuration package into dedicated modules
