+++ library/clients/pubchem.py
@@ -17,7 +17,8 @@ import requests
 from cachetools import TTLCache
 from requests import Session
 
-from ..config import ApiCfg, PubChemCfg, RetryCfg, session_with_retry
+from ..config.models import ApiCfg, PubChemCfg, RetryCfg
+from ..config.runtime import session_with_retry
 from ..common.log import logger
 from ..common.rate_limiter import get_limiter, sleep
