+++ library/config/__init__.py
@@ -1,0 +1,38 @@
+"""High level configuration interface for ChEMBL data acquisition."""
+
+from __future__ import annotations
+
+from . import models as models
+from .loader import (
+    ConfigError,
+    _absolutise_path_value,
+    _mask_secrets,
+    _serialize_paths,
+    build_alias_map,
+    load_config,
+    print_config,
+)
+from .models import *  # noqa: F401,F403
+from .runtime import (
+    configure_rate_limits,
+    crossref_session,
+    ensure_dirs,
+    openalex_session,
+    session_with_retry,
+)
+
+__all__ = [
+    *models.__all__,  # type: ignore[name-defined]
+    "ConfigError",
+    "_absolutise_path_value",
+    "_mask_secrets",
+    "_serialize_paths",
+    "build_alias_map",
+    "load_config",
+    "print_config",
+    "configure_rate_limits",
+    "ensure_dirs",
+    "session_with_retry",
+    "openalex_session",
+    "crossref_session",
+]
